refactor: 로또 구입 관련 로직 분리
Payment
1. 구입금액 유효성 검사
2. 로또번호 발급
